KAZMIN-90: changed type from to double

+++ modules/web/src/kz/uco/tsadv/web/modules/personal/absencebalance/AbsenceBalanceBrowse.java
@@ -175,10 +175,10 @@ public class AbsenceBalanceBrowse extends EditableFrame {
         newBalance.setPersonGroup(personGroup);
         newBalance.setDateFrom(dateFrom);
         newBalance.setDateTo(dateTo);
-        newBalance.setBalanceDays(balanceDays);
+        newBalance.setBalanceDays((double) balanceDays);
 //        newBalance.setAdditionalBalanceDays(additionalBalanceDays);
-        newBalance.setDaysSpent(annualDaysSpent);
-        newBalance.setExtraDaysSpent(extraDaysSpent);
+        newBalance.setDaysSpent((double) annualDaysSpent);
+        newBalance.setExtraDaysSpent((double) extraDaysSpent);
 //        newBalance.setExtraDaysLeft(extraDaysLeft);
         if (daysLeft < 0) {
             daysLeft = 0;
